+++ cmt/src/kernels/library.m
@@ -63,7 +63,7 @@ MT_API_AVAILABLE(mt_macos(10.11), mt_ios(8.0))
 void
 mtLibraryFunctionNames(MtLibrary *lib, size_t* count, const char** names) {
   NSArray<NSString*> *_names = [(id<MTLLibrary>)lib functionNames];
-  int n = [_names count];
+  NSInteger n = [_names count];
 
   if (*count == 0) {
     *count = n;
